+++ demo/demo-symfony6/src/Form/ChangePasswordType.php
@@ -22,14 +22,12 @@ class ChangePasswordType extends AbstractType
                 'first_options' => [
                     'label' => 'New Password',
                     'constraints' => [
-                        new NotBlank([
-                            'message' => 'Please enter a password',
-                        ]),
-                        new Length([
-                            'min' => 6,
-                            'minMessage' => 'Your password should be at least {{ limit }} characters',
-                            'max' => 4096,
-                        ]),
+                        new NotBlank(message: 'Please enter a password'),
+                        new Length(
+                            min: 6,
+                            max: 4096,
+                            minMessage: 'Your password should be at least {{ limit }} characters'
+                        ),
                     ],
                 ],
                 'second_options' => [