+++ demo/demo-symfony7/src/Form/UserType.php
@@ -27,14 +27,12 @@ class UserType extends AbstractType
                 'mapped' => false,
                 'required' => !$options['is_edit'],
                 'constraints' => [
-                    new NotBlank([
-                        'message' => 'Please enter a password',
-                    ]),
-                    new Length([
-                        'min' => 6,
-                        'minMessage' => 'Your password should be at least {{ limit }} characters',
-                        'max' => 4096,
-                    ]),
+                    new NotBlank(message: 'Please enter a password'),
+                    new Length(
+                        min: 6,
+                        max: 4096,
+                        minMessage: 'Your password should be at least {{ limit }} characters'
+                    ),
                 ],
             ]);
     }